apparmor: add a transition flags to override NO_NEW_PRIVS
NO_NEW_PRIVS attempts to restricts all apparmor transitions that
would grant additional privileges to a process. Practically,
determining if a transition from one profile to another results in
a strict subset of rules is expensive and complicated, so apparmor
only allows transitions to a child profile.

We'd like to add an additional permission flag, but xindex doesn't
have a free bit to use. For now, use the same value as AA_X_UNSAFE
and change this when we have a different packing strategy for
xindex.

This change permits the the following transition under NNPs even if
"bar" isn't a child profile of "foo":

	foo px -> bar, # Unsafe transitions override NO_NEW_PRIVS

Future work will add a new flag instead of piggybacking off of an
unsafe transtion:

	foo PNx -> bar, # Allow transition despite NO_NEW_PRIVS

Signed-off-by: Eric Chiang <ericchiang@google.com>