security/apparmor: fix matching on presence of extended attributes
AppArmor recently added the ability for profiles to match extended
attributes, with the intent of targeting "security.ima" and
"security.evm" to differentiate between sign and unsigned files.

The current implementation uses a path glob to match the extended
attribute value. To require the presence of a extended attribute,
profiles supply a wildcard:

    # Match any file with the "security.apparmor" attribute
    profile test /** xattrs=(security.apparmor="**") {
        # ...
    }

However, the glob matching implementation is intended for file paths and
doesn't handle null characters correctly. It's currently impossible to
write a profile that targets IMA and EVM attributes, since the
signatures can contain a null byte.

Add the ability for AppArmor to check the presence of an extended
attribute, and not its value. This fixes the profile matching allowing
profiles conditional on EVM and IMA signatures:

    profile signed_binary /** xattrs=(security.evm security.ima) {
        # ...
    }

A modified apparmor_parser and associated regression tests to exercise
this fix can be found at:

    https://gitlab.com/ericchiang/apparmor/commits/parser-xattrs-keys

Signed-off-by: Eric Chiang <ericchiang@google.com>
CC: stable@vger.kernel.org